+++ src/components/ItemAccordion/ItemAccordion.js
@@ -1,8 +1,9 @@
-
+import PropTypes from 'prop-types';
 import React from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { Field } from 'react-final-form';
 
+import { NumberGeneratorModalButton } from '@folio/service-interaction';
 import {
   Accordion,
   Row,
@@ -18,15 +19,28 @@ import {
   useData,
   useOptions,
 } from '../../hooks';
+import {
+  BARCODE_SETTING,
+  NUMBER_GENERATOR_OPTIONS_ON_EDITABLE,
+  NUMBER_GENERATOR_OPTIONS_ON_NOT_EDITABLE,
+} from '../../consts';
 
 import { validateBarcode } from '../../util';
 
-const ItemAccordion = () => {
-  const { materialTypes, loanTypes } = useData();
+const ItemAccordion = ({ change }) => {
+  const {
+    materialTypes,
+    loanTypes,
+    numberGeneratorSettings,
+  } = useData();
+
   const materialTypeOptions = useOptions(materialTypes, 'id', 'name');
   const permanentLoanTypeOptions = useOptions(loanTypes, 'id', 'name');
   const { formatMessage } = useIntl();
   const { okapi } = useStripes();
+
+  const isBarcodeDisabled = numberGeneratorSettings?.[BARCODE_SETTING] === NUMBER_GENERATOR_OPTIONS_ON_NOT_EDITABLE;
+  const showNumberGeneratorForBarcode = isBarcodeDisabled || numberGeneratorSettings?.[BARCODE_SETTING] === NUMBER_GENERATOR_OPTIONS_ON_EDITABLE;
 
   return (
     <Accordion
@@ -36,6 +50,7 @@ const ItemAccordion = () => {
       <Row>
         <Col sm={4}>
           <Field
+            disabled={isBarcodeDisabled}
             label={<FormattedMessage id="ui-plugin-create-inventory-records.barcode" />}
             name="item.barcode"
             id="barcode"
@@ -43,6 +58,22 @@ const ItemAccordion = () => {
             validate={validateBarcode(okapi)}
             fullWidth
           />
+          <Row>
+            {showNumberGeneratorForBarcode &&
+              <Col xs={12}>
+                <NumberGeneratorModalButton
+                  buttonLabel={<FormattedMessage id="ui-plugin-create-inventory-records.numberGenerator.generateBarcode" />}
+                  callback={(generated) => change('item.barcode', generated)}
+                  id="inventoryItemBarcode"
+                  generateButtonLabel={<FormattedMessage id="ui-plugin-create-inventory-records.numberGenerator.generateBarcode" />}
+                  generator="inventory_itemBarcode"
+                  modalProps={{
+                    label: <FormattedMessage id="ui-plugin-create-inventory-records.numberGenerator.generateBarcode" />
+                  }}
+                />
+              </Col>
+            }
+          </Row>
         </Col>
         <Col sm={4}>
           <Field
@@ -73,6 +104,10 @@ const ItemAccordion = () => {
       <ElectronicAccessFields />
     </Accordion>
   );
+};
+
+ItemAccordion.propTypes = {
+  change: PropTypes.func.isRequired,
 };
 
 export default ItemAccordion;